requirements.txt
requests
selenium
newspaper3k
selenium-stealth
beautifulsoup4